+++ belief_propagation/src/lib.rs
@@ -33,15 +33,15 @@ use crate::{factor_graph::CTFactorGraph, zero_lookahead_belief_propagation::{run
 /// `[node_name, posterior_probability_1, node_category]`
 pub fn zero_lookahead_bp(
     graph: String,
-    alpha: f64,
-    beta: f64,
+    alpha: f32,
+    beta: f32,
     regularized: bool,
-    prior: f64,
+    prior: f32,
     max_iter: Option<u32>,
-    tolerance: Option<f64>
+    tolerance: Option<f32>
 ) -> String {
     let max_iter: u32 = max_iter.unwrap_or(10000);
-    let tolerance: f64 = tolerance.unwrap_or(0.006);
+    let tolerance: f32 = tolerance.unwrap_or(0.006);
     
     run_belief_propagation(graph, alpha, beta, regularized, prior, max_iter, tolerance).unwrap()
 }
@@ -67,6 +67,7 @@ pub fn load_factor_graph(
     graph: String,
 ) -> Result<Vec<CTFactorGraph>, Box<dyn std::error::Error>> {
     let mut ct_factor_graph = CTFactorGraph::from_graphml(&graph)?;
+
     ct_factor_graph.add_ct_nodes();
 
     let ct_factor_graphs: Vec<CTFactorGraph> = ct_factor_graph.connected_components();
@@ -92,22 +93,22 @@ pub fn load_factor_graph(
 /// `[node_name, posterior_probability_1, node_category]`
 pub fn zero_lookahead_bp_from_graph(
     graphs: &mut Vec<CTFactorGraph>,
-    alpha: f64,
-    beta: f64,
+    alpha: f32,
+    beta: f32,
     regularized: bool,
-    prior: f64,
+    prior: f32,
     max_iter: Option<u32>,
-    tolerance: Option<f64>
+    tolerance: Option<f32>
 ) -> Result<String, Box<dyn std::error::Error>> {
     let max_iter: u32 = max_iter.unwrap_or(10000);
-    let tolerance: f64 = tolerance.unwrap_or(0.006);
+    let tolerance: f32 = tolerance.unwrap_or(0.006);
 
     for ct_factor_graph in graphs.iter_mut() {
         ct_factor_graph.fill_in_factors(alpha, beta, regularized);
         ct_factor_graph.fill_in_priors(prior);
     }
 
-    let results: Vec<(String, Vec<f64>)> = calibrate_all_subgraphs(
+    let results: Vec<(String, Vec<f32>)> = calibrate_all_subgraphs(
         graphs,
         max_iter,
         tolerance
